Move ES from contrib to core. Add optimization test functions in ES. Add rank transformation. Update setup scripts.
Former-commit-id: 0aa9d5c389fc3e14ba4e38842c8f863bbf0944b6

+++ lagom/core/transform/normalize.py
@@ -37,4 +37,4 @@ class Normalize(BaseTransform):
         # Normalize data into [0, 1]
         out = (x - min_val)/(max_val - min_val + self.eps)
         
-        return out
+        return out.astype(np.float32)
